# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BankCards API",
    "description": "REST API для управления банковскими картами",
    "contact": {
      "name": "Kitoglav",
      "email": "vlad_lanayev@vk.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/api/cards/block/{cardId}": {
      "post": {
        "tags": [
          "Карты"
        ],
        "summary": "Заблокировать карту",
        "description": "Блокирует карту пользователя",
        "operationId": "blockCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Карта заблокирована",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Нет доступа к карте",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/cards": {
      "get": {
        "tags": [
          "Администрирование карт"
        ],
        "summary": "Получить все карты",
        "description": "Возвращает список всех карт в системе с пагинацией",
        "operationId": "getAllCards",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Параметры пагинации",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получен список карт",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Администрирование карт"
        ],
        "summary": "Создать новую карту",
        "description": "Создает новую банковскую карту для пользователя",
        "operationId": "createCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Карта создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные карты",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/cards/block/{cardId}": {
      "post": {
        "tags": [
          "Администрирование карт"
        ],
        "summary": "Заблокировать карту",
        "description": "Блокирует карту в системе",
        "operationId": "blockCard_1",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Карта заблокирована",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/cards/activate/{cardId}": {
      "post": {
        "tags": [
          "Администрирование карт"
        ],
        "summary": "Активировать карту",
        "description": "Активирует заблокированную карту",
        "operationId": "activateCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Карта активирована",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cards": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получить все карты пользователя",
        "description": "Возвращает список всех карт пользователя с пагинацией",
        "operationId": "getCards",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Параметры пагинации",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получен список карт",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{cardId}": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получить карту по ID",
        "description": "Возвращает информацию о конкретной карте пользователя",
        "operationId": "getCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Карта найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Нет доступа к карте",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/cards/{cardId}": {
      "get": {
        "tags": [
          "Администрирование карт"
        ],
        "summary": "Получить карту по ID",
        "description": "Возвращает информацию о карте по ее идентификатору",
        "operationId": "getCard_1",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Карта найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Администрирование карт"
        ],
        "summary": "Удалить карту",
        "description": "Удаляет карту из системы",
        "operationId": "deleteCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Карта удалена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "204": {
            "description": "Карта не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "hiddenNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "expiryDate": {
            "type": "string",
            "format": "date"
          },
          "balance": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateCardRequest": {
        "type": "object",
        "description": "Данные для создания карты",
        "properties": {
          "cardNumber": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelCardResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "description": "Введите JWT токен без префикса 'Bearer'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}